update scatterbrain cache client, make it more generic (could handle either webGL or WebGPU easily)
