+++ packages/scatterbrain/src/index.ts
@@ -1,8 +1,8 @@
+export { buildScatterbrainCacheClient } from './cache-client';
+export { getVisibleItems, loadDataset as loadScatterbrainDataset } from './dataset';
 export {
     buildRenderFrameFn as buildScatterbrainRenderFn,
-    buildScatterbrainCacheClient,
     setCategoricalLookupTableValues,
-    updateCategoricalValue,
+    updateCategoricalValue
 } from './renderer';
 export * from './types';
-export { getVisibleItems, loadDataset as loadScatterbrainDataset } from './dataset';
